requirements.txt
# Core
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Database
mysqlclient==2.2.7
psycopg2-binary==2.9.10

# API Documentation
drf-yasg==1.21.10
PyYAML==6.0.2
uritemplate==4.2.0

# Environment and Configuration
django-environ==0.10.0
python-decouple==3.8

# Chapa Integration
requests==2.31.0
python-dotenv==1.0.0

# Asynchronous Tasks
celery==5.5.3
redis==6.2.0
django-redis==5.4.0

# Development
# debug-toolbar temporarily removed due to Python 3.12 compatibility issues
whitenoise==6.9.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0

# Chapa Dependencies (included in requirements but managed by other packages)
amqp==5.3.1
billiard==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
sqlparse==0.5.3
typing-extensions==4.12.2
vine==5.1.0
wcwidth==0.2.13
